% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{aov_auto}
\alias{aov_auto}
\title{Automatically Compute Parametric or Non-parametric t-test or ANOVA}
\usage{
aov_auto(
  df,
  type = c("all", "params"),
  na.rm_type = c("cols", "rows"),
  norm_test = c("all", "res"),
  p.adjust.methods = "bonferroni",
  quiet = F
)
}
\arguments{
\item{df}{A data.frame that has factor and numeric variables}

\item{type}{(Character) If \code{type = "params"}: It will only compute shapiro and levene test}

\item{na.rm_type}{(Character) Specify how to remove \code{NA} values i.e., by rows ("rows") or columns ("cols").}

\item{norm_test}{(Character)
\itemize{
\item \strong{"all"}: perform normality test on numeric variables: \code{shapiro.test(num_var)}.
\item \strong{"res"}: perform normality test on residual of linear model: \code{shapiro.test(residual(lm(num_var ~ fct_var)))}.
}}

\item{p.adjust.methods}{(Character) P-value adjustment method pass to p.adjust()}

\item{quiet}{(logical) Give message about removed variables or not?}
}
\value{
A data.frame
}
\description{
Automatically perform t-test or ANOVA according to the number of levels of factors.
Assumptions were checked prior to choosing parametric or non-parametric methods.
All combination of factor and numeric variables in data.frame will be computed.
}
\examples{
aov_auto(iris)

}
