% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_helper.R
\name{complete_rate}
\alias{complete_rate}
\title{Calculate Completeness of Data}
\usage{
complete_rate(x, negate = FALSE)
}
\arguments{
\item{x}{list or data.frame}

\item{negate}{\code{TRUE} to return proportion of missing data}
}
\value{
A vector of proportion of completeness (not missing)
}
\description{
Calculate proportion of completeness (no \code{NA}) of the element in the data frame or list
}
\examples{
df <- data.frame(x = c(1, NA, 2),
                 y = c(NA, "a", "b"),
                 z = 1:3)

complete_rate(df)
}
