% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_var_table}
\alias{count_var_table}
\title{Count variable and display to table}
\usage{
count_var_table(x, var, sort = TRUE, name = "Count", add_total = TRUE)
}
\arguments{
\item{x}{A data.frame}

\item{var}{(Quote or Unquoted) Variable to count}

\item{sort}{Output sorted or not}

\item{name}{Name of column that store the "Count".}

\item{add_total}{Add "Total" at last row or not}
}
\value{
A data.frame
}
\description{
This function count 1 variable and display the output as count and percentage.
Also add "Total" at last row.
}
\examples{
count_var_table(iris, Species)
}
