% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_vars}
\alias{count_vars}
\title{Count Each Variables to List of Data Frames}
\usage{
count_vars(x, ..., sort = TRUE, name = NULL, prop = FALSE)
}
\arguments{
\item{x}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{Variables to count (one level); If not supply, count all variables.}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{name}{The name of the new column in the output.}

\item{prop}{If \code{TRUE} calculate proprotion for each count.}
}
\value{
A named list of data.frame
}
\description{
Count all or many variables in a data.frame (depth 1 level) and display the results as list of data.frame.
This function intended for interactive use.
}
\examples{
# Count All variables
count_vars(iris)
# Count Specified variables
count_vars(iris, Species, Sepal.Length)
}
