% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_vars_plot}
\alias{count_vars_plot}
\title{Count variable for plotting}
\usage{
count_vars_plot(
  x,
  ...,
  var_desc = TRUE,
  label_number = c("percent", "count", "both"),
  lump_prop = 0,
  other_level = "Other"
)
}
\arguments{
\item{x}{A data.frame}

\item{...}{(Quoted or Unquoted) Variables to count}

\item{var_desc}{If \code{TRUE} levels of the variable will be reorder in descending order according to
count of that variable}

\item{label_number}{Construction of \code{labels} column in the output
\itemize{
\item "percent": use \verb{\%} (eg. 60\%)
\item "count": count as integer (eg. 6)
\item "both": use \verb{\%} and count in parenthesis  (eg. 60\% (6))
}}

\item{lump_prop}{Proportion that lump levels to "Other" (passed to \code{fct_lump} at \code{prop} argument)}

\item{other_level}{Name of the "Other" level}
}
\value{
A data.frame with "n", "prop", "pc", "labels" column
}
\description{
This function count variables, do \code{fct_reorder} by that count, extract percentage and construct \code{labels} column
ready to be plotted.
}
\examples{
count_vars_plot(iris, Species)
}
