% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_vars_plot_facet}
\alias{count_vars_plot_facet}
\title{Count Variables to Plot Bar Graph with Facet}
\usage{
count_vars_plot_facet(
  x,
  var,
  var_facet,
  label_number = c("percent", "count", "both")
)
}
\arguments{
\item{x}{A data.frame}

\item{var}{(Quoted or Unquoted) Variable to count}

\item{var_facet}{(Quoted or Unquoted) Variable intended to get facet wrapped (grouping variable)}

\item{label_number}{Construction of \code{labels} column in the output
\itemize{
\item "percent": use \verb{\%} (eg. 60\%)
\item "count": count as integer (eg. 6)
\item "both": use \verb{\%} and count in parenthesis  (eg. 60\% (6))
}}
}
\value{
A data.frame
}
\description{
A Helper that count variable in each group of \code{var_facet} do extract percentage and construct \code{labels} column
ready to be plotted.
}
\examples{
count_vars_plot_facet(mtcars, cyl, gear)
}
