% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_vars_sep.R
\name{count_vars_sep}
\alias{count_vars_sep}
\title{Separate and Count Variables to DF}
\usage{
count_vars_sep(data, vars = NULL, sep = ",", convert = FALSE, sort = TRUE)
}
\arguments{
\item{data}{A Data Frame}

\item{vars}{(Character) A Character of variables to count. If \code{NULL}, count all variables in the data frame.}

\item{sep}{A separator of texts}

\item{convert}{If \code{TRUE} will automatically run \code{\link[=type.convert]{type.convert()}} on the key column. This is useful if the column types are actually numeric, integer, or logical.}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}
}
\value{
a named list of data frames (named as counted variable names)
}
\description{
Separate texts in a variables using a separator \code{sep} into rows, trim white space,
counts each variables, and return list of counted data frames.
}
\examples{
df <- data.frame(
  x = 1:3,
  y = c("a", "d,e,f", "g,h"),
  z = c("1", "2, 3, 4", "5, 6, 7")
 )
count_vars_sep(df, sep = ",")
}
