% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoder.R
\name{encoder}
\alias{encoder}
\title{Encode Vector}
\usage{
encoder(x, match, encode = match, nomatch_na = TRUE)
}
\arguments{
\item{x}{(vector) Input data to be matched.}

\item{match}{(vector) Value to be matched against \code{x}'s element}

\item{encode}{(vector) Encoding vector same length as \code{match}}

\item{nomatch_na}{(Logical) Control behavior of non-matching elements, one of:
\itemize{
\item \strong{\code{TRUE}:} (default) return \code{NA} for non-matching elements.
\item \strong{\code{FALSE}:} return original elements of \code{x}.
}}
}
\value{
Encoded vector
}
\description{
Match and Encode vector
}
\examples{
encoder(c("a","b","d"), c("a","b","c"), c("A","B","C"))
encoder(c("a","b","d"), c("a","b","c"), c("A","B","C"), nomatch_na = FALSE)
}
