% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_helper.R
\name{fill_na}
\alias{fill_na}
\title{Fill Missing values from another vector}
\usage{
fill_na(x, fill)
}
\arguments{
\item{x}{Vector to fill \code{NA}}

\item{fill}{A Vector whose elements will be retrieved to fill in \code{x}}
}
\value{
A Vector with original element from \code{x} and filled \code{NA} from \code{fill}
}
\description{
Fill \code{NA}s values from \code{x} using values from \code{fill} that in the same location.
Vector \code{x} and \code{fill} must have the same length.
}
\examples{

v1 <- c(NA, NA, "a", NA, "b") # Trust v1
v2 <- c(NA, "c", "d","k", NA)
fill_na(v1, v2)

}
