% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbx-package.R
\docType{package}
\name{lbx-package}
\alias{lbx}
\alias{lbx-package}
\title{lbx: My Data Exploration Package}
\description{
Wrapper of data manipulating and plotting function.
}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
