% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot Counted Bar Graph}
\usage{
plot_bar(
  data,
  var,
  var_desc = TRUE,
  lump_prop = 0,
  other_level = "Other",
  alpha = 0.8,
  stat = c("percent", "prop", "count"),
  label_geom = c("label", "label_repel", "text", "text_repel"),
  label_number = c("percent", "count", "both"),
  ...
)
}
\arguments{
\item{data}{A data.frame}

\item{var}{(Quote or Unquote) Categorical variables to plot}

\item{var_desc}{If \code{TRUE} levels of the variable will be reorder in descending order according to
count of that variable}

\item{lump_prop}{Proportion that lump levels to "Other" (passed to \code{fct_lump} at \code{prop} argument)}

\item{other_level}{Name of the "Other" level}

\item{alpha}{An alpha of \code{geom_col}}

\item{stat}{(A Character) How y-axis computed ?
\itemize{
\item "percent": as percentage with \verb{\%}
\item "prop": as proportion
\item "count": as count
}}

\item{label_geom}{(A Character) Indicate function of label geom
\itemize{
\item "label": \code{geom_label}
\item "label_repel": \code{ggrepel::geom_label_repel}
\item "text": \code{geom_text}
\item "text_repel": \code{ggrepel::geom_text_repel}
}}

\item{label_number}{(A Character) How to display the label?
\itemize{
\item "percent": use \verb{\%} (eg. 60\%)
\item "count": count as integer (eg. 6)
\item "both": use \verb{\%} and count in parenthesis  (eg. 60\% (6))
}}

\item{...}{Passed to selected function of \code{label_geom}}
}
\value{
A Plot class \code{gg} and \code{ggplot}
}
\description{
Plot counted of 1 variable using bar graph
}
\examples{
plot_bar(mtcars, cyl)
}
