% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count.R
\name{plot_bars_facet2D}
\alias{plot_bars_facet2D}
\title{Plot Bar Graph with Facet Wrap}
\usage{
plot_bars_facet2D(
  data,
  var,
  var_facet,
  stat = c("percent", "prop", "count"),
  label_geom = c("label", "label_repel", "text", "text_repel"),
  label_number = c("percent", "count", "both"),
  ...
)
}
\arguments{
\item{data}{A data.frame}

\item{var}{(Quoted or Unquoted) Variable to count in each facet}

\item{var_facet}{(Quoted or Unquoted) Variable intended to get facet wrapped (grouping variable)}

\item{stat}{(A Character) How y-axis computed ?
\itemize{
\item "percent": as percentage with \verb{\%}
\item "prop": as proportion
\item "count": as count
}}

\item{label_geom}{(A Character) Indicate function of label geom
\itemize{
\item "label": \code{geom_label}
\item "label_repel": \code{ggrepel::geom_label_repel}
\item "text": \code{geom_text}
\item "text_repel": \code{ggrepel::geom_text_repel}
}}

\item{label_number}{(A Character) How to display the label?
\itemize{
\item "percent": use \verb{\%} (eg. 60\%)
\item "count": count as integer (eg. 6)
\item "both": use \verb{\%} and count in parenthesis  (eg. 60\% (6))
}}

\item{...}{Passed to selected function of \code{label_geom}}
}
\value{
A Plot class \code{gg} and \code{ggplot}
}
\description{
Plot count of 2 variables (2D) into bar graph with facet wrap.
}
\examples{
plot_bars_facet2D(mtcars, cyl, gear, color = "white")
}
