% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count.R
\name{plot_pie}
\alias{plot_pie}
\title{Plot Pie Chart}
\usage{
plot_pie(
  data = NULL,
  var,
  lump_prop = 0,
  other_level = "Other",
  polar_start = 0,
  polar_direction = 1,
  polar_clip = "on",
  label_geom = c("label", "label_repel", "text", "text_repel"),
  label_number = c("percent", "count", "both"),
  color_label = "black",
  ...
)
}
\arguments{
\item{data}{A data.frame}

\item{var}{(Quote or Unquote) Categorical variables to plot}

\item{lump_prop}{Proportion that lump levels to "Other" (passed to \code{fct_lump} at \code{prop} argument)}

\item{other_level}{Name of the "Other" level}

\item{polar_start}{The \code{start} arg of the \code{coord_polar()}}

\item{polar_direction}{The \code{direction} arg of the \code{coord_polar()}: 1 for clockwise rotation from high-to-low count.}

\item{polar_clip}{The \code{clip} arg of the \code{coord_polar()}}

\item{label_geom}{(A Character) Indicate function of label geom
\itemize{
\item "label": \code{geom_label}
\item "label_repel": \code{ggrepel::geom_label_repel}
\item "text": \code{geom_text}
\item "text_repel": \code{ggrepel::geom_text_repel}
}}

\item{label_number}{(A Character) How to display the label?
\itemize{
\item "percent": use \verb{\%} (eg. 60\%)
\item "count": count as integer (eg. 6)
\item "both": use \verb{\%} and count in parenthesis  (eg. 60\% (6))
}}

\item{color_label}{(A Character) Color for label}

\item{...}{Passed to selected function of \code{label_geom}}
}
\value{
A Plot class \code{gg} and \code{ggplot}
}
\description{
High level constructor of pie chart using \code{ggplot2} to display count of 1 variable.
}
\examples{
plot_pie(mtcars, cyl)
}
