% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_cnd.R
\name{print_messages}
\alias{print_messages}
\title{Print Character Vector to Messages}
\usage{
print_messages(x, sep = "\\n", prefix = "", suffix = "")
}
\arguments{
\item{x}{A character vector}

\item{sep}{Separator between each character elements}

\item{prefix}{Prefix of each character elements}

\item{suffix}{Suffix of each character elements}
}
\value{
A message
}
\description{
Print Character Vector to Messages
}
