% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_helper.R
\name{remove_na_rows}
\alias{remove_na_rows}
\title{Remove Rows that has NAs}
\usage{
remove_na_rows(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
A data.frame with complete cases
}
\description{
Remove rows that has any \code{NA} values
}
\examples{

df <- data.frame(x = c(1, 2, NA),
                 y = c(1, NA, 3))
remove_na_rows(df)
}
